Ninja: Add compile rules only for languages that are actually compiled
In `WriteLanguageRules` we collect all languages used for source files
in a target.  However, this only needs to include sources that are
actually going to be compiled into object files.  No object file build
statements will be generated for other sources.

This avoids generating language compile rules for source files that
are not compiled due to being marked as `HEADER_FILE_ONLY`.

Issue: #16373