VS: Disable project upgrade prompt on VS 2013 and above
When opening projects in Visual Studio that specify an older toolset
version, a prompt is displayed the first time asking the user whether to
upgrade the projects.  This is meant for project files that are
maintained manually and updated through the IDE.  For CMake-generated
projects this does not make sense, so add content to tell VS not to
upgrade.