cmake: Comment why we ignore the include-what-you-use return code
The include-what-you-use tool always returns non-zero to indicate that
it did not actually produce an object file as Clang would from the same
command line.  Add a comment explaining that this is why we ignore its
return code.  Also update our `pseudo_iwyu` test suite tool to always
exit with an error too.