FortranCInterface: Fix support for VS with Intel toolset
When using a Visual Studio generator with an Intel toolset, such as

    -T "Intel C++ Compiler XE 14.0"

the generated FortranCInterface mangling detection project may fail to
build due to `devenv` not working with the `/project ALL_BUILD` option.
This seems to be a bug in `devenv` or the Intel VS integration.  Work
around the problem by building with `/project FortranCInterface`
instead.  We only need to build this executable and its dependencies
within the detection test project anyway.

Fixes: #16519