AndroidTestUtilities: ignore exit status of ls
adb used to suffer from a bug where `adb shell` would return 0 irrespective
the exit status of the underlying process that it executed.  This is now
fixed for newer versions of Android (N+).

As a result, attempting to `adb shell ls` a nonexistent file can now result
in a non-zero exit status.  Updating check_device_file_exists to ignore the
exit status of `adb shell` fixes this bug for newer Android devices.