FindHDF5: Fix search for language-specific header file
Refactoring in commit v3.6.0-rc1~85^2 (HDF5: Refactor the use of
compiler wrappers, 2016-04-04) converted code of the form

    if(${LANGUAGE} MATCHES ...)

to

    if(LANGUAGE MATCHES ...)

However, `LANGUAGE` is a foreach() loop variable and not a normal
variable so auto-dereference does not occur.  Restore the explicit `${}`
syntax and use the new name of the loop variable that has changed since
then too.

Fixes: #16651