FindVulkan: Fix for SDK versions < 1.0.42 on 32-bit Windows
Before Vulkan 1.0.42, the SDK would add the Bin directory to `PATH`,
which was confusing the 32-bit search on Windows.  Avoid such confusion
by ignoring the `PATH` when looking for the 32-bit version.

Vulkan 1.0.42 fixed the whole problem by moving the libraries into Lib,
Lib32, so this is strictly a compatibility fix for old SDKs.