FindHDF5: set HDF5_<lang>_INCLUDE_DIRS in all cases
In the case that we're using `find_path` to find the include directory,
only the `_DIR` variable was set, while in all other cases only the
`_DIRS` variable was set.  Set the `_DIRS` variable in all cases (while
keeping `_DIR` as result of `find_path`).  Also document the `_DIRS`
variable.

Fixes: #16694