update-kwsys: Disable custom gitattributes during import
The KWSys `.gitattributes` file now defines custom attributes.
Git does not support these in a subdirectory where we are about
to place the content, so comment them out during import.
This also ensures that the `our-c-style` attribute on KWSys
sources does not activate the `format.clang-format` attribute
on the third-party import branch.