Tests: Add option to customize LinkInterfaceLoop timeout
This test has a timeout in case CMake gets into an infinite loop.  The
default of 90 seconds should be plenty of time for the test to run
correctly since it does not actually do a build.  However, busy machines
that run lots of tests in parallel may need a longer timeout.  Give them
an option to extend it.