FindPythonInterp: Add `-32` and `-64` registry entry variants
As of 3.6.1 Python (only version tested in this case), installing 32-bit
version on 64-bit windows has a slightly different registry key path.
`-32` is appended to the end of the version number in the path.  Also
added a few more registry path guesses based on 64-bit version of
Python.

Also the module now checks if the host is WIN32 instead of the target,
which is required if you are cross compiling.