FindwxWidgets: Verify existence of libraries reported by wx-config
When the Ubuntu package `libwxgtk3-dev` is not installed, FindwxWidgets
succeeds even the required libraries are not actually found. The reason
for that is that `wx-config` does not perform any kind of check and uses
hard-coded values.  This affects e.g. KiCAD build process and makes it
fail with a confusing error message [1].  The solution is to check for
presence of every library file obtained from `wx-config`.

[1] https://bugs.launchpad.net/kicad/+bug/1630020