VS2017: Verify Windows 8.1 SDK before using it
The detection logic added by commit v3.8.0-rc2~14^2 (VS2017: If Win 8.1
SDK is not available, use Win 10 SDK, 2017-02-20) was incomplete.  It is
possible for the Win 8.1 SDK registry entry to exist, and even the
directory, but the header files to not actually be installed.  Teach
`cmGlobalVisualStudio15Generator::IsWin81SDKInstalled` to verify that
the `windows.h` header actually exists in the SDK directory.  We do this
in `cmGlobalVisualStudio14Generator::GetWindows10SDKVersion` for the
Windows 10 SDK already.

Fixes: #16811