bindexplib: Skip symbols containing a dot (.)
Symbols including a dot are not valid and result in a `LNK1242` error
when trying to create a library from the def file.  Such symbols happen
to be in object files when using PGI Fortran on Windows and compiling
with debug symbols enabled.  Those symbols do not need to be exported.