FindICU: Search for packaged data files
The files Makefile.inc and pkgdata.inc are installed by ICU to
<libdir>/icu and on some distributions to <libdir>/<arch>/icu.
These are required by ICU tools like pkgdata, and so are needed
by other package build systems at build time, in addition to the
libraries and executables already searched for.

Search for the files and create cache entries for them, as done
for executables.